fix: Fix to use SliverList in SessionsTable
